Addition of Frases table and Styling
This change includes Tablero changes to accomodate Frases instead of Description.
It also includes a lot of styling for all the administration forms and for the Tablero. They are now all responsive and fit a phone.

If a card does not have a Frase it will use as text the Name of the Card.